refactor: rename CLI app
