+++ deployer/__main__.py
@@ -2,7 +2,7 @@
 # registered into the app
 
 from . import commands, dev_commands  # noqa: F401
-from .cli_app import app
+from .app import app
 from .utils.jsonnet import validate_jsonnet_version
