+++ deployer/commands/plan_upgrade/jobs.py
@@ -4,7 +4,7 @@ import os
 import typer
 from ruamel.yaml import YAML
 
-from deployer.cli_app import CONTINUOUS_DEPLOYMENT, app
+from deployer.app import CONTINUOUS_DEPLOYMENT, app
 from deployer.utils.file_acquisition import REPO_ROOT_PATH, get_all_cluster_yaml_files
 from deployer.utils.rendering import create_markdown_comment, print_colour
