+++ deployer/dev_commands/exec/cost_monitoring/cost_monitoring_app.py
@@ -9,7 +9,7 @@ import requests
 import typer
 from yarl import URL
 
-from deployer.cli_app import exec_app
+from deployer.app import exec_app
 from deployer.commands.validate.config import cluster_config as validate_cluster_config
 from deployer.dev_commands.exec.aws.aws_app import setup_aws_sts_env
 from deployer.infra_components.cluster import Cluster
