+++ deployer/dev_commands/generate/cryptnono_config/__init__.py
@@ -2,7 +2,7 @@ import shutil
 import subprocess
 from pathlib import Path
 
-from deployer.cli_app import generate_app
+from deployer.app import generate_app
 from deployer.utils.file_acquisition import REPO_ROOT_PATH
 
 HERE = Path(__file__).parent
