+++ deployer/dev_commands/generate/dedicated_cluster/dedicated_cluster_app.py
@@ -6,7 +6,7 @@ under the `generate` sub-command of the deployer.
 
 import typer
 
-from deployer.cli_app import generate_app
+from deployer.app import generate_app
 
 dedicated_cluster_app = typer.Typer(pretty_exceptions_show_locals=False)
 generate_app.add_typer(
