+++ deployer/dev_commands/update/eksctl.py
@@ -1,6 +1,6 @@
 import typer
 
-from deployer.cli_app import update_app
+from deployer.app import update_app
 from deployer.dev_commands.generate.dedicated_cluster.aws import generate_eksctl
 from deployer.infra_components.cluster import Cluster
 from deployer.utils.rendering import print_colour
